Convert Nav and ApplicationBase to functional components
This PR is fairly long, but it converts the two components into its functional
counterpart.

With the `Nav` we converted to `Navigation`that exports a `NavigationProvider`
component and a `NavigationContext`, the latter is handy if we wanted to
access visit, remote, pageKey, and navigateTo to help users build a custom
`<Link>` component or if they want to use the functions without passing it
from parent to child.

With `Application`, we removed `ApplicationBase`. In doing so, we've simplified
the component to the point that if someone wanted fine grain control of building
the redux `Provider` or `Navigation`, they can just refer to the source code and
build their own using the same helper methods in `lib/index.ts`.

We also updated the docs.