Simplify generators
This moves the install generator and scaffold generator under a superglue namespace.

With the install generator, now we have `rails g superglue:install`, making use of
rails generators also gives us the ability to add options to the install.

With the scaffold generator, now we have `rails g superglue:scaffold`. This replaces
the previous clunky scaffold that had to disable templates.