Remove ownprops
`ownProps` is a way to pass additional props to the page component when navigating, but its usage
runs counter to the way pages are restored. It's essentially temporary and doesn't play well
with a refresh. I think search params on the pageKey might be a better approach, and similar to
ember's philosophy of the url as the reflection of application state. For now, lets remove
the functionality.