Add jsx rendering defaults
This improves the developer experience by adding rendering defaults for superglue views.
Superglue typically requires 3 templates.

```
app/views/
  posts/
    index.html.erb
    index.jsx
    index.json.props
  users/
    index.html.erb
    index.jsx
    index.json.props
```
In most cases `index.html.erb` is the same, so the immediate thought is to remove it. This PR
makes the following possible:

```
app/views
  application/
    superglue.html.erb
  posts/
    index.jsx
    index.json.props
  users/
    index.jsx
    index.json.props
```

In the above scenario, we can now render a common template using `superglue_template` in the controller:

```
class PostsController < ApplicationController
  before_action :use_jsx_rendering_defaults
  superglue_template "application/superglue"
end
```

This PR also adds a custom resolver that allows for the following scenario:

```
app/views
  application/
    superglue.html.erb
  posts/
    index.jsx
  users/
    index.jsx
```

This case is great for projects that have their own API and just want to use
superglue for the rails like routing and take advantage of the UJS helpers. In
that case, they can skip using `props_template`. Note that we still need the
`app/views/layout/application.json.props` that gets generated.