Remove remove_to
We're removing this because a removal may potentially break state. I'm not dealing with DOM
here, Superglue works with composable data fragments that can reference other data fragments,
if one got deleted, it may cause state issues. Instead I encourage to NOT delete things,
and to just add a `deleted_at` attribute. This way the frontend stays stable and can
hide or show based on that attribute.