Remove before_to and after_to
In turbo rails this is about insert a DOM element before/after another DOM element. In
superglue, the closest is inserting into a list before/after a target element. Not quite
equivalent, but also not as useful since we can sort client side. I think i'll
remove this for now.