Remove render: false
In turbo render: false is to used to render nothing, but still use the
attributes passed to mutate a dom element's attributes. This isn't necessary
in Superglue. The closes to this feature is options, which allow passing options
to the key actions (append, replace, etc.) for additional functionality
(eg. unique pushes)