Fixes inline rendering
This change fixes an issue with inline rendering a props_template
in combination with a JSX file. Since both jsx files and props file
need to be side by side, using the virtual path of the jsx file would
be valid enough. I'll need a change in props_template to consider the variable.